+++ client/presentation/src/main/java/com/andone/memorip/presentation/placedetail/component/ContrastAwareText.kt
@@ -1,17 +1,19 @@
 package com.andone.memorip.presentation.placedetail.component
 
-import android.R.attr.height
 import android.graphics.Bitmap
-import android.graphics.Color
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import com.andone.memorip.presentation.placedetail.component.Constants.cropRate
+import com.andone.memorip.presentation.placedetail.component.Constants.luminusThreshold
 import com.andone.memorip.presentation.theme.MemoripTheme
+import com.andone.memorip.presentation.util.calculateLuminusAverage
+import com.andone.memorip.presentation.util.cropBottomArea
 
 private object Constants {
     val cropRate = 0.2f
+    val luminusThreshold = 128
 }
 
 @Composable
@@ -21,9 +23,9 @@ fun ContrastAwareText(
     style: TextStyle
 ) {
     val textColor = if (image != null) {
-        val cropImage = cropTopArea(bitmap = image)
-        val luminusValue = calculateLuminusAverage(bitmap = cropImage)
-        if (luminusValue > 128) MemoripTheme.colors.black
+        val luminusValue = image.cropBottomArea(ratio = cropRate)
+            .calculateLuminusAverage()
+        if (luminusValue > luminusThreshold) MemoripTheme.colors.black
         else MemoripTheme.colors.white
     } else {
         MemoripTheme.colors.white
@@ -33,36 +35,6 @@ fun ContrastAwareText(
         text = text,
         style = style,
         color = textColor
-    )
-}
-
-private fun calculateLuminusAverage(bitmap: Bitmap): Double {
-    var sum = 0.0
-    val width = bitmap.width
-    val height = bitmap.height
-    val pixels = IntArray(width * height)
-
-    bitmap.getPixels(pixels, 0, width, 0, 0, width, height)
-
-    pixels.forEach { pixel ->
-        val r = Color.red(pixel)
-        val g = Color.green(pixel)
-        val b = Color.blue(pixel)
-        sum += 0.299 * r + 0.587 * g + 0.114 * b
-    }
-
-    return sum / pixels.size
-}
-
-private fun cropTopArea(bitmap: Bitmap, ratio: Float = cropRate): Bitmap {
-    val endY = (bitmap.height * ratio).toInt()
-    val startY = bitmap.height - endY
-    return Bitmap.createBitmap(
-        bitmap,
-        0,
-        startY,
-        bitmap.width,
-        endY
     )
 }
